Fixed Branching Logic and other issues
The problems we had in the lab we caused by some poor logic on how
read/write indexes and data were being assigned. Fixed those up so our
Register data look good. There is one aspect that seems to work but does
not make me happy. The PC that is sent to the control module is the next
program counter. Which means we do not need to add for a No-Branch_PC.
But it means we need to decrement the FetchData.PC by 1 to have correct
branch-relative results.